requirements.txt
Django>=4.2,<5
djangorestframework>=3.14
djangorestframework-simplejwt
PyJWT>=2.8
mysqlclient>=2.1
python-dotenv>=1.0
django-environ>=0.11
